use StreamingHttpResponse instead of channels
