--- django_eventstream/routing.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import consumers
-
-urlpatterns = [
-    path("", consumers.EventsConsumer.as_asgi()),
-]
