+++ examples/chat/requirements.txt
@@ -1,7 +1,5 @@
-asgiref==3.5.2
-channels==3.0.5
-daphne==3.0.2
-dj-database-url==0.5.0
-Django==3.2.14
+daphne==4.1.0
+dj-database-url==2.1.0
+Django==5.0.2
 django-dotenv==1.4.2
-django-grip==3.4.0
+django-grip==3.5.0
